+++ simauto/intersection.py
@@ -86,7 +86,7 @@ class IntersectionEnv(gym.Env):
 
         self.mean_waits = []
 
-        self.episode_end_threshold = 100 # in numbers of cars passed
+        self.episode_end_threshold = 1000000 # in numbers of cars passed
 
         self.action_space = spaces.Discrete(6)
         self.observation_space = gym.spaces.Dict({
